Strong, Massive, Resist orbs ig
Welp they're functional and also were pain to add in so h

+++ battle/data/MaskEnemy.java
@@ -9,7 +9,7 @@ import java.util.TreeSet;
 
 public interface MaskEnemy extends MaskEntity {
 
-	double getDrop();
+	int getDrop();
 
 	@Override
 	Enemy getPack();
